+++ src/cn/salesuite/saf/imagecache/BitmapConnection.java
@@ -6,7 +6,12 @@ package cn.salesuite.saf.imagecache;
 import java.io.IOException;
 import java.io.InputStream;
 
+import org.apache.http.HttpResponse;
+import org.apache.http.client.methods.HttpGet;
+
+import cn.salesuite.saf.download.AndroidHttpClient;
 import cn.salesuite.saf.http.rest.RestClient;
+import cn.salesuite.saf.utils.SAFUtils;
 
 /**
  * @author Tony Shen
@@ -17,9 +22,24 @@ class BitmapConnection {
 		T run(InputStream stream);
 	}
 
-	public <T> T readStream(String urlString,Runnable<T> runnable) {		
-		RestClient client = RestClient.get(urlString).useCaches(true);
-		InputStream inputStream = new PatchInputStream(client.stream());
+	public <T> T readStream(String urlString,Runnable<T> runnable) {
+		InputStream inputStream = null;
+		if (SAFUtils.isLOrHigher()) {
+			AndroidHttpClient client = AndroidHttpClient.newInstance("BitmapConnection");
+			HttpGet httpGet = new HttpGet(urlString);
+			
+			HttpResponse response;
+			try {
+				response = client.execute(httpGet);
+				inputStream = new PatchInputStream(response.getEntity().getContent());
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		} else {
+			RestClient client = RestClient.get(urlString).useCaches(true);
+			inputStream = new PatchInputStream(client.stream());
+		}
+		
 		try {
 			return runnable.run(inputStream);
 		} finally {